2020 Days 1 and 2: Removed unnecessary PRINT with FORMAT
Thanks to u/symbolicprocessor on r/adventofcode for this tip: when
T is specified as the first argument of FORMAT, the formatted
string will be printed to stdout and the return value will be NIL,
which is why I was seeing NIL's being printed when the top-level
statements were evaluated.

I also switched to using path literals (#p"") for file paths.  It
would be nice if there were a reader macro for raw strings so that
I wouldn't have to double-escape backslashes, but at least CL-PPCRE
is super nice for text extraction.