Added shorthand lambda reader macro
I learned about reader macros!  Then I promptly used them to implement
something that's probably horrifying!  :D

In short, I defined a reader macro that allows lambdas to be written
like Smalltalk blocks, being delimited with [] and with arguments
referred to positionally as %1, %2, etc. and the &REST as %&.  If only
a single argument is present, % can be used instead of a numbered
argument.  Arguments are automatically declared ignorable, so positional
arguments can be omitted freely.  If the first value in the body of the
lambda is a sexp, then the body will be surrounded with PROGN.

Note that with the current implementation, shorthand lambdas cannot be
nested.  I consider this to be a feature rather than a flaw, similar to
Clojure's restriction on its shorthand anonymous functions.  It would
theoretically be possible to avoid this restriction by GENSYM-ing away
the lambda parameters, but that's honestly too much work to enable an
antipattern.

Hopefully this shorthand will encourage me to start using lambdas more.
I've updated my past solutions to use the new syntax, and I'm already
substantially more comfortable using it than the verbose LAMBDA offered
by default in Common Lisp.

Note that the following lines of Emacs Lisp are required to ensure that
automatic indentation works correctly for this new syntax:
>> (modify-syntax-entry ?[ "(]" lisp-mode-syntax-table)
>> (modify-syntax-entry ?] ")[" lisp-mode-syntax-table)

This feature and its implementation were inspired both by
http://www.chadbraunduin.com/2011/07/common-lisp-reader-macros.html
and the rutils package (https://github.com/vseloved/rutils/).  Thanks
to them for the great work!