2020 Day 4: Solution Complete
This was a bit more challenging of a problem than the previous days,
both due to the slightly increased difficulty of parsing the input
file (stubbornly attempting to do all of the parsing in a single LOOP
was not a good idea) and my weird implicit assumption that
CL-PPCRE:SCAN would match the entire line (it does not).

Regardless, I think that my solution is pretty clean, especially with
the shorthand lambdas for validator functions.  I also started to learn
Alexandria (mainly for WHEN-LET(*)) and the cl-str library, both of
which I hope will continue to serve me well on future challenges and
which seem to be good additions to my personal Common Lisp "standard
library".