2020 Day 4: Added number length validation
Added some validation to the numeric fields to ensure that they are
not preceded or followed by unknown junk characters.  This wasn't
necessary for today's puzzle, but I feel like a future day will
involve passport-checking again (CID, anyone?) so it's best that my
validation routine is robust as possible from the outset.