2020 Day 5: Added solution
Today's problem was substantially simpler than yesterday's, if for
no other reason than that it didn't require too much input parsing.
Nevertheless, it allowed me to continue developing my skills with
the LOOP macro: at this point, I think I've used almost every
directive in an Advent of Code solution at some time or another.

My solution for Part 2 is the simple arithmetic approach.  I know
that you can use the bit-twiddling expression:
  >> (XOR over list) XOR (XOR over (range list))
To determine a missing integer in a sequence as well, but that
honestly wasn't necessary for today's solution since the sum of all
of the seat ID's wouldn't even overflow a 32-bit integer.  I went
with the arithmetic solution largely because I liked the code a bit
better, although I've confirmed that both will work in Lisp.