2020 Day 6: Added submission
This problem was honestly extremely simple.  Two set operations and
an addition reduction (via LOOP -> SUM) was all that was necessary
to solve both parts.  A fun trick that I realised was that I could
simplify the blank-line input parsing by virtually inserting a blank
line at the end of the file by setting the EOF value for READ-LINE
to "".