2020 Day 8: Added solution
Added my solution for 2020 Day 8.  This solution is horrifically
over-engineered, but it served as a great introduction to CLOS and
should hopefully be fairly useful on future days if this year
becomes Intcode Part 2.  Plus, I like writing VM's/emulators and
this was a great opportunity to try writing one in Lisp.  Lisp is
actually more comfortable for this problem domain than I was
expecting with some additions from the Alexandria and CL-STR
libraries, and CLOS is a pretty nice OOP system.  Hopefully I'll
be able to grow the switch-case soon :D