2020 Day 9: Added solution
Added my solution for 2020 Day 9.  I'm not sure I'm fully satisfied
with the prolixity of my solution (why doesn't ALEXANDRIA:MAP-COMBINATIONS
actually return the mappings?), but I'm reasonably happy with Part 2.
In addition, I'm pleased with the constant-time sliding window
implementation (HEAD-SEQ . TAIL-SEQ), which is definitely a trick
to remember for future days.