2020 Day 10: Added solution
This was definitely a shorter problem, although it served as a great
testbed for my new solution framework.  It turns out that having a
general structure to the solution when starting out was really
helpful!  Part 2 was basically a classic "dump everything in a hash
table" DP problem, although it allowed me to transfer that general
problem structure over to Lisp, which was fairly simple using a COND
statement.