Added CI workflow for acceptance tests
After an incredible amount of mucking about with GitHub actions, this
repo now includes a CI workflow that will ensure that all acceptance
tests are passing.  The signalling mechanism for failure isn't the
best (disabling the SBCL debugger then causing fiveam to enter the
debugger when a test fails), but it shows the failure and is thus
good enough for my uses.