2020 Day 11: Added solution
This was a slightly longer problem, although cellular automata are
always fun.  Part 1 was relatively straightforward, and I was able
to adapt my solution for Part 2 somewhat well by parameterising the
tolerance and neighbour checking function throughout the automaton
runner.  However, Part 2 proved to be a pain because of a smattering
of off-by-one errors and my ignorance as to how the THEREIS and
NEVER clauses for LOOP work.  Speaking of which, the real MVP of
today's problem is the LOOP macro, which drives functionally the
entire solution: there's no time for recursion when your solution
is already pretty slow.