2020 Day 12: Added solution
I'm pretty proud of my solution for this one: realising that I could
parameterise the movement function to allow the "direction" in Part 1
to simply be a complex number waypoint was a huge epiphany that
allowed me to delete a lot of repetitive code, and the use of complex
numbers to represent coordinates made the maths wonderfully concise.
Also, there may actually be more documentation than code for this
solution, but I think that's warranted.