2020 Day 15: Added solution
Here's my solution for Day 15 of this year's challenge.  This was an
astonishingly simple problem: a pretty straightforward linear
sequence for which no optimisations can be performed.  Switching the
data store from a hash table to an array helped the performance a
fair deal: it turns out that integer hashing functions and conflicts
can indeed have an affect on runtime.