2020 all days: Switched to Package Inferred System
In order to avoid future identifier conflicts, I've switched my
project structure from everything being in a single namespace to
using package-inferred-system, in which each file is its own project.
This is probably the way that projects should be structured in the
first place, although I'm somewhat surprised that it took me this
long to discover it.  The acceptance tests should now be passing
again!