2020 Day 16: Added solution
Here's my solution for Day 16 of this year's challenge.  This was a
pretty enjoyable challenge: the input parsing didn't take as long as
I was expecting it to, and writing a SAT solver was fun.  The only
real stumbling block was that there were tickets with an error rate
of 0 that were still invalid...this was extremely frustrating to
debug, but I was able to figure out the issue courtesy of a meme on
/r/adventofcode.