2020 Day 17: Added solution
Here's my solution for Day 17 of this year's challenge.  On the
surface, this was a pretty simple modification of Day 11, but then
I decided to overengineer everything.  The Conway solver now works
for any number of dimensions >= 2, although confusingly it exhibits
a memory leak from MAPCAR in NEXT-STATE.  I've spent enough time
trying to debug this and think I'll just attribute it to unoptimised
memory usage for now.