2020 Day 18: Added solution
Here's my solution for Day 18 of this year's challenge!  This was a
fun problem: I knew exactly what Part 2 was going to be when I saw
Part 1, so I used Shunting-Yard from the outset.  I think my
implementation of the algorithm is pretty clean, being a "pure
functional" approach with ne'er a SETF in sight.