2020 Day 19: Added solution
Here's my solution for Day 19 of this year's challenge.  Apparently
the goal of today was to write a parser for a simple context-free
grammar, which I wasn't aware of but is definitely pretty neat.  My
solution generates the lengths of all matching prefixes for the
ruleset, which works acceptably well and has fairly reasonable
performance.